refactor: move create_player_test to utils
