+++ app/tests/services/test_strokes_service.py
@@ -2,21 +2,10 @@ import uuid
 
 from sqlmodel.ext.asyncio.session import AsyncSession
 
-from app.models.player import Player, PlayerCreate
+from app.models.player import PlayerCreate
 from app.services.players_and_strokes_service import PlayersAndStrokesService
-from app.services.players_service import PlayersService
 from app.services.strokes_service import StrokesService
-
-
-async def create_player(session: AsyncSession, user_public_id: uuid.UUID) -> Player:
-    user_public_id_str = str(user_public_id)
-    telegram_id = 10103030
-    player_service = PlayersService()
-    player_create = PlayerCreate(
-        user_public_id=user_public_id_str, telegram_id=telegram_id
-    )
-    player = await player_service.create_player(session, player_create)
-    return player
+from app.tests.utils.utils import create_player
 
 
 async def test_create_strokes_defaults_to_beginner_base_level(
