+++ app/tests/utils/utils.py
@@ -1,7 +1,12 @@
 import random
 import string
+import uuid
+
+from sqlmodel.ext.asyncio.session import AsyncSession
 
 from app.core.config import settings
+from app.models.player import Player, PlayerCreate
+from app.repository.players_repository import PlayersRepository
 
 
 def random_lower_string() -> str:
@@ -11,3 +16,14 @@ def random_lower_string() -> str:
 def get_x_api_key_header() -> dict[str, str]:
     headers = {"x-api-key": f"{settings.API_KEY}"}
     return headers
+
+
+async def create_player(session: AsyncSession, user_public_id: uuid.UUID) -> Player:
+    user_public_id_str = str(user_public_id)
+    telegram_id = 10103030
+    repo = PlayersRepository(session)
+    player_create = PlayerCreate(
+        user_public_id=user_public_id_str, telegram_id=telegram_id
+    )
+    player = await repo.create_player(player_create)
+    return player
